agent_demo/requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
protobuf==3.7.0
pycryptodome==3.7.3
requests==2.21.0
six==1.12.0
urllib3==1.24.1
